+++ src/main/java/community/whatever/onembackendjava/common/exception/ErrorCode.java
@@ -5,10 +5,14 @@ import org.springframework.http.HttpStatus;
 public enum ErrorCode {
 
 
+    // Authorization
+    EXPIRED_SHORTEN_URL(HttpStatus.UNAUTHORIZED, "만료된 단축 URL입니다."),
 
+    // Validation
     INVALID_URL_FORMAT(HttpStatus.BAD_REQUEST, "잘못된 URL 형식입니다."),
     BLOCKED_URL(HttpStatus.FORBIDDEN, "사용 불가능한 URL입니다."),
 
+    // NotFound
     NOT_FOUND_SHORTEN_URL(HttpStatus.NOT_FOUND, "단축 URL을 찾을 수 없습니다.");
