+++ src/main/java/community/whatever/onembackendjava/common/exception/custom/UnAuthorizedException.java
@@ -1,2 +1,10 @@
-package community.whatever.onembackendjava.common.exception.custom;public class UnAuthorizedException {
+package community.whatever.onembackendjava.common.exception.custom;
+
+import community.whatever.onembackendjava.common.exception.ErrorCode;
+
+public class UnAuthorizedException extends BusinessException{
+
+    public UnAuthorizedException(ErrorCode errorCode) {
+        super(errorCode);
+    }
 }
