+++ src/main/java/community/whatever/onembackendjava/urlshorten/service/UrlShortenService.java
@@ -2,6 +2,7 @@ package community.whatever.onembackendjava.urlshorten.service;
 
 import community.whatever.onembackendjava.common.exception.ErrorCode;
 import community.whatever.onembackendjava.common.exception.custom.NotFoundException;
+import community.whatever.onembackendjava.common.exception.custom.UnAuthorizedException;
 import community.whatever.onembackendjava.urlshorten.component.UrlShortenValidator;
 import community.whatever.onembackendjava.urlshorten.component.UrlShortener;
 import community.whatever.onembackendjava.urlshorten.domain.UrlShorten;
@@ -36,7 +37,7 @@ public class UrlShortenService {
             .orElseThrow(() -> new NotFoundException(ErrorCode.NOT_FOUND_SHORTEN_URL));
 
         if (urlShorten.getExpiredAt().isBefore(LocalDateTime.now())) {
-            throw new IllegalArgumentException("만료된 URL입니다.");
+            throw new UnAuthorizedException(ErrorCode.EXPIRED_SHORTEN_URL);
         }
 
         return urlShorten.getOriginUrl();
