+++ src/test/java/community/whatever/onembackendjava/urlshorten/controller/UrlShortenControllerTest.java
@@ -1,0 +1,16 @@
+package community.whatever.onembackendjava.urlshorten.controller;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+class UrlShortenControllerTest {
+
+    @Test
+    void shorten_url을_생성한다() {
+
+    }
+
+
+
+}
