+++ src/test/java/community/whatever/onembackendjava/urlshorten/service/UrlShortenServiceTest.java
@@ -1,11 +1,16 @@
-package community.whatever.onembackendjava.application;
+package community.whatever.onembackendjava.urlshorten.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import community.whatever.onembackendjava.common.exception.ErrorCode;
 import community.whatever.onembackendjava.common.exception.custom.NotFoundException;
+import community.whatever.onembackendjava.common.exception.custom.UnAuthorizedException;
 import community.whatever.onembackendjava.common.exception.custom.ValidationException;
+import community.whatever.onembackendjava.urlshorten.domain.UrlShorten;
+import community.whatever.onembackendjava.urlshorten.repository.UrlShortenRepository;
 import community.whatever.onembackendjava.urlshorten.service.UrlShortenService;
+import java.time.LocalDateTime;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -16,6 +21,9 @@ class UrlShortenServiceTest {
     @Autowired
     private UrlShortenService urlShortenService;
 
+    @Autowired
+    private UrlShortenRepository urlShortenRepository;
+
     @Test
     void shorten_url을_생성하고_조회한다() {
         String expectedOriginUrl = "https://www.google.com";
@@ -24,14 +32,13 @@ class UrlShortenServiceTest {
 
         assertThat(originUrl).isEqualTo(expectedOriginUrl);
     }
-
     @Test
     void 잘못된_url일_경우_예외가_발생한다() {
-        String originUrl = "ftp://www.google.com";
+        String originUrl = "htp://www.google.com";
 
         assertThatThrownBy(() -> urlShortenService.createShortenUrl(originUrl))
             .isInstanceOf(ValidationException.class)
-            .hasMessage("잘못된 URL 형식입니다.");
+            .hasMessage(ErrorCode.INVALID_URL_FORMAT.getMessage());
     }
 
     @Test
@@ -40,7 +47,7 @@ class UrlShortenServiceTest {
 
         assertThatThrownBy(() -> urlShortenService.createShortenUrl(originUrl))
             .isInstanceOf(ValidationException.class)
-            .hasMessage("사용 불가능한 URL입니다.");
+            .hasMessage(ErrorCode.BLOCKED_URL.getMessage());
     }
 
     @Test
@@ -49,8 +56,19 @@ class UrlShortenServiceTest {
 
         assertThatThrownBy(() -> urlShortenService.getOriginUrlByShortenUrlKey(nonExistingKey))
             .isInstanceOf(NotFoundException.class)
-            .hasMessage("단축 URL을 찾을 수 없습니다.");
+            .hasMessage(ErrorCode.NOT_FOUND_SHORTEN_URL.getMessage());
 
+    }
+
+    @Test
+    void 만료된_shorten_url을_조회하면_예외가_발생한다() {
+        String expiredKey = "expiredKey";
+        UrlShorten expiredUrl = new UrlShorten("https://google.com", expiredKey, LocalDateTime.now().minusMinutes(1));
+        urlShortenRepository.save(expiredUrl);
+
+        assertThatThrownBy(() -> urlShortenService.getOriginUrlByShortenUrlKey(expiredKey))
+            .isInstanceOf(UnAuthorizedException.class)
+            .hasMessage(ErrorCode.EXPIRED_SHORTEN_URL.getMessage());
     }
 
 }
